Separate libusb and libserial implementations
